Island: Add MongoEventRepository to DI container
